Changing migrations to fix duplication of tables being created

+++ 2024_Performance_Hub/database/migrations/2024_12_12_135526_create_musician_performance_table.php
@@ -11,12 +11,14 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::create('musician_performance', function (Blueprint $table) {
-            $table->id();
-            $table->foreignId('musician_id')->constrained()->onDelete('cascade');
-            $table->foreignId('performance_id')->constrained()->onDelete('cascade');
-            $table->timestamps();
-        });
+        if (!Schema::hasTable('musician_performance')) {
+            Schema::create('musician_performance', function (Blueprint $table) {
+                $table->id();
+                $table->foreignId('musician_id')->constrained()->onDelete('cascade');
+                $table->foreignId('performance_id')->constrained()->onDelete('cascade');
+                $table->timestamps();
+            });
+        }
     }
 
     /**
